Add MacBookAir 6,2 support
Detect the machine type via dmi and add the proper sysfs files to handle the
backlight brightness level.

We rely on the fact that the backlight driver in use is mba6x_bl. It should work
without it, but the stock intel backlight kernel driver sets the screen brigthness at
~5% or ~95% after suspend, rendering the screen too dark or too light until you
reboot.

Had to add the sysfs file paths for the mba6x_bl driver _before_ the intel ones
because the "discovery" logic just loops the array, and even when the mba6x_bl driver
is loaded the intel sysfs files were used to set the backlight level.

Please note that USB product IDs other than the EUR one are placeholders since I
don't have access to other devices. That should be fixed.

Signed-off-by: L. Alberto Giménez <agimenez@sysvalve.es>